refactor email sending to own module
