+++ index.js
@@ -105,4 +105,4 @@ app.get("*", (req, res) => {
 });
 
 // initialise server
-app.listen(PORT, () => console.log(`Server listening on port ${PORT}!`))
+app.listen(PORT, () => console.log(`Server listening on port ${PORT}!`));
